refactor(ntt): 优化NTT实现并改进数值类型处理
重构NTT模块，将i32类型改为i16以提升性能，简化乘法运算
更新barrett_reduce和montgomery_reduce函数以匹配新类型
调整ZETAS常量为i16类型，优化内存使用

+++ src/poly.rs
@@ -31,7 +31,7 @@ impl Poly {
     #[inline(always)]
     pub fn reduce(&mut self) {
         for coeff in self.coeffs.iter_mut() {
-            *coeff = barrett_reduce(*coeff as i64) as i16;
+            *coeff = barrett_reduce(*coeff as i32);
         }
     }
